server: FEATURE: socket communication permissions
Use configure's --with-user and --with-group options to limit netopeer
usage to USER that can run the netopeer-server and GROUP of users
that are allowed to access it via netopeer-agent.

It is required that USER is root or a member of GROUP.

By default, both these variables are set by configure to root value.