server: BUGFIX: Netopeer transAPI module loading
Netopeer transAPI module is static and loaded automatically. It is
an implicit netopeer-server module that cannot be (un)loaded, so
it is not supposed to be present in its NETCONF configuration data.
This commit removes the default config data withg info about
loading Netopeer module. It caused loading the module for the second
time and therefore printing error messages.