toaster: BUGFIX: Deklaring local variable static prevents race condition causing infinite sleep.
Function rpc_make_toast launch function make_toast in a new thread
passing doneness as a parameter.
Unfortunately, doneness is a local parameter when the new thread uses
it... it as been scratch by "something else".
The new thread will then sleep for an unpredictable amount of time.

Fixes issue 12