server: BUGFIX: changing ownership on install
When server is installed, it changes ownership for some files
to allow access for specified USER/GROUP. iBut this is possible only
when installation is performed by root. If not (e.g. testing install
into DESTDIR), install process crashed. Now, the process of changing
ownership is skipped in this case and applied only when make is run
by root.