manager: BUGFIX: adding first Netopeer module
When Netopeer does not contain any transAPI module spec in its datastore,
but the datastore file correctly exists (it is empty), netopeer-manager
have to create correct content of the configuration datastore. So far it
was only searching for the specific element where the newly added module
is going to be placed, but in the described case, the parent element does
not exist.