server: CHANGE: rename intermediate files generated by configure
change suffix of intermediate files generated by configure that
possibly include ${prefix} variable. They must be again processed
by Make to expand this variable. Previously, the suffix was .out,
but it is recognized by make as target of the implicit rule and
using it caused circular dependency. Now, it is renamed to .tmp
and the files generated by configure were placed into Makefile
as requirements in rule definitions.