cfgsystem: BUGFIX: configuration data validity
When creating running configuration data on transapi_init(), the
list nodes were created invalid - the key elements of the list item
must be placed first in order how they are specified in the key
declaration. We placed them after other elements due to the code
optimization - it was a bug.