server: FEATURE: recover datastore locks
When the server unexpectedly terminates and the datastores are locked,
it failed to correctly load because the initial copy-config failed and
netopeer transAPI module was unable to load other transAPI modules.

Now, during the initial start of the server, it breaks the datastore
locks before the init of the module.