turing: BUGFIX: do not print empty cells in status
cells with no symbol are not allowed in the status data. The previous
commit solves only the end of tape, but some rules can also rewrite the
cell symbol with empty symbol ('\0') and make the repo invalid this way.
In such a case, the length of the tape is still the same, but the cells
with an empty symbol are not visible in the status data.