server: BUGFIX: installing sshd_config from RPM
In case of installing netopeer-server from RPM, sshd_config is taken
from the target system as part of the post install steps. However,
the make was still copying (or trying to copy) the sshd_config from the
building machine into the RPM package.

We add --with-rpm option to the configure which tells Make to do not
install sshd_config from the local system into the netopeer's config path.
The option is used by the RPM spec file.