server: FEATURE: added converter and yang parser
Added files needed for quick testing of converter inside restconf
server. This is a WIP and will later be removed in favor of
having these files in a separate library (hopefuly). For now a
dependency exists between the agent and converter in the form of
a connection being passed from agent to converter.